Resharper suggestions
