+++ src/RedisObjectCache/RedisCache.cs
@@ -1,11 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Linq;
 using System.Runtime.Caching;
-using System.Text;
 using System.Threading;
-using System.Threading.Tasks;
 using StackExchange.Redis;
 
 namespace RedisObjectCache
