+++ src/RedisObjectCache/RedisCacheEntry.cs
@@ -1,10 +1,5 @@
 ﻿using System;
-using System.CodeDom;
-using System.Collections.Generic;
-using System.Linq;
 using System.Runtime.Caching;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace RedisObjectCache
 {
